speed up Play urls and the web player process (#2635)
* add urls to playlist in one go
* don't re-add to the playlist when it's there
* allow 0 duration for time when you get items
* remove track_num from stream_url
* add codec to stream_url
* give every media type a non-static play_url
* remove format checks
* set the filetype in the url

+++ play/index.php
@@ -35,8 +35,8 @@ ob_end_clean();
 
 /* These parameters had better come in on the url. */
 $uid          = scrub_in($_REQUEST['uid']);
-$object_id    = scrub_in($_REQUEST['oid']);
-$sid          = scrub_in($_REQUEST['ssid']);
+$object_id    = (int) scrub_in($_REQUEST['oid']);
+$sid          = (int) scrub_in($_REQUEST['ssid']);
 $type         = (string) scrub_in(filter_input(INPUT_GET, 'type', FILTER_SANITIZE_SPECIAL_CHARS));
 $cache        = scrub_in($_REQUEST['cache']);
 $format       = scrub_in($_REQUEST['format']);
